Create 'siteHasSubscription' selector.
